refactor(harness): remove the surfaces the llm-router supersedes
Deleted: the local provider-router decide() library, the provider registry
(harness::provider::register/resolve/list) and its refresh-on-config bridge,
the models-catalog module (models::list/get/supports/reconcile + scope
'models'), per-provider Phase-1 complete functions, and the compaction
stream collector. The catalog Model type moves to types/model.ts, wire-
aligned with the router (provider-side extras are optional — the router
catalog doesn't persist them). The harness entry is permissions-only; the
harness worker now depends on llm-router instead of models-catalog.
