+++ harness/package.json
@@ -27,7 +27,6 @@
     "dev:session": "tsx src/session/main.ts",
     "dev:llm-budget": "tsx src/llm-budget/main.ts",
     "dev:hook-fanout": "tsx src/hook-fanout/main.ts",
-    "dev:models-catalog": "tsx src/models-catalog/main.ts",
     "dev:provider-anthropic": "tsx src/provider-anthropic/main.ts",
     "dev:provider-openai": "tsx src/provider-openai/main.ts",
     "dev:provider-kimi": "tsx src/provider-kimi/main.ts",
@@ -44,7 +43,6 @@
     "iii-session": "./dist/session/main.js",
     "iii-llm-budget": "./dist/llm-budget/main.js",
     "iii-hook-fanout": "./dist/hook-fanout/main.js",
-    "iii-models-catalog": "./dist/models-catalog/main.js",
     "iii-provider-anthropic": "./dist/provider-anthropic/main.js",
     "iii-provider-openai": "./dist/provider-openai/main.js",
     "iii-provider-kimi": "./dist/provider-kimi/main.js",
