+++ harness/src/harness/iii.worker.yaml
@@ -22,7 +22,7 @@ dependencies:
   iii-sandbox: "^0.11.0"
   iii-directory: "^0.5.1"
   turn-orchestrator: "^0.2.0"
-  models-catalog: "^0.2.0"
+  llm-router: "^0.1.0"
   shell: "^0.3.0"
   provider-anthropic: "^0.2.0"
   provider-openai: "^0.2.0"
