--- harness/src/models-catalog/iii.worker.yaml
@@ -1,17 +1,0 @@
-iii: v1
-name: models-catalog
-language: node
-deploy: binary
-manifest: package.json
-bin: iii-models-catalog
-description: Model capabilities knowledge base on the iii bus (models::*). State-backed catalog written by provider discovery via models::reconcile.
-
-runtime:
-  kind: node
-
-scripts:
-  install: pnpm install
-  start: node ./dist/models-catalog/main.js
-
-dependencies:
-  iii-state: "^0.11.0"
