+++ harness/src/turn-orchestrator/provisioning/ports.ts
@@ -2,7 +2,7 @@
  * Typed dependency ports for provisioning.
  */
 
-import type { Model } from '../../models-catalog/types.js';
+import type { Model } from '../../types/model.js';
 import type { ISdk } from '../../runtime/iii.js';
 import { logger } from '../../runtime/otel.js';
 import { getCatalogModel } from '../../runtime/output-tokens.js';
