+++ harness/src/turn-orchestrator/state.ts
@@ -4,7 +4,7 @@
  * State is cleared on deploy, so records are read back typed (no re-parsing).
  */
 
-import type { Model } from '../models-catalog/types.js';
+import type { Model } from '../types/model.js';
 import type { AssistantMessage, FunctionResultMessage } from '../types/agent-message.js';
 import type { ExecutedCall, FunctionBatchWork, PreparedCall } from './function-execute/types.js';
